+++ sdk/keyvault/Azure.Security.KeyVault.Secrets/ChangeLog.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 4.0.0-preview.4
+
+### Breaking changes
+
+- `SecretBase` has been renamed to `SecretProperties`.
+- `Secret` and `DeletedSecret` no longer extend `SecretProperties`, but instead contain a `SecretProperties`.
+- `SecretProperties.Vault` has been renamed to `SecretProperties.VaultUri`.
+
 ## 4.0.0-preview.1 (2019-06-28)
 Version 4.0.0-preview.1 is the first preview of our efforts to create a user-friendly client library for Azure Key Vault. For more information about
 preview releases of other Azure SDK libraries, please visit